medico

+++ Projeto Consultorio Nivel 2/Models/Consulta.cs
@@ -6,7 +6,7 @@
         public DateTime DataConsulta { get; set; } = DateTime.MinValue;
         public string Descricao { get; set; } = string.Empty;
         public string PrescricaoMedica { get; set; } = string.Empty;
-        public Decimal ValorConsulta { get; set; } = 0;
+        public double ValorConsulta { get; set; } = 0;
         public int IdMedico { get; set; }
         public int IdPaciente { get; set; }
         public Medico Medico { get; set; }
